reverted to previous version

+++ cfm/cfm-code/EmNNModel.h
@@ -3,9 +3,9 @@
 #
 # EM.h
 #
-# Description: 	Class to apply Expectation Maximization algorithm to derive
+# Description: 	Class to apply Expectation Maximization algorithm to derive 
 #				model parameters when using a neural net for theta.
-#				- all identical to linear model, except params are NNParam and gradient
+#				- all identical to linear model, except params are NNParam and gradient 
 #				  computation is different.
 #					E-step: IPFP or equivalent.
 #					M-step: Gradient Ascent
@@ -24,34 +24,33 @@
 
 #include "EmModel.h"
 #include "NNParam.h"
-#include <cstddef>
 
 class EmNNModel : public EmModel {
 public:
-	// Constructor
-	// Note: To include the group in the status filename, include _GRP_ in the name
-	EmNNModel(config_t *a_cfg, FeatureCalculator *an_fc, std::string &a_status_filename,
-	          std::string initial_params_filename = "");
+    //Constructor
+    //Note: To include the group in the status filename, include _GRP_ in the name
+    EmNNModel(config_t *a_cfg, FeatureCalculator *an_fc, std::string &a_status_filename,
+              std::string initial_params_filename = "");
 
-	// This is public so the test can access it....there must be a better way?
-	int computeAndAccumulateGradient(double *grads, int mol_idx, MolData &mol_data, suft_counts_t &suft,
-	                                 int sampling_method, unsigned int energy) override;
+    //This is public so the test can access it....there must be a better way?
+    int computeAndAccumulateGradient(float *grads, int mol_idx, MolData &mol_data, suft_counts_t &suft,
+                                     int sampling_method, unsigned int energy) override;
 
-	void collectUsedIdx(MolData &mol_data, std::set<unsigned int> &used_idxs, unsigned int energy) override;
+    void collectUsedIdx(MolData &mol_data, std::set<unsigned int> &used_idxs, unsigned int energy) override;
 
-	double computeLogLikelihoodLoss(int molidx, MolData &moldata, suft_counts_t &suft, unsigned int energy) override;
+    double computeLogLikelihoodLoss(int molidx, MolData &moldata, suft_counts_t &suft, unsigned int energy) override;
 
-	double getRegularizationTerm(unsigned int energy) override;
+    double getRegularizationTerm(unsigned int energy) override;
 
-	void updateGradientForRegularizationTerm(double *grads, unsigned int energy) override;
+    void updateGradientForRegularizationTerm(float *grads, unsigned int energy) override;
 
-	void writeParamsToFile(std::string &filename) override;
+    void writeParamsToFile(std::string &filename) override;
 
 private:
-	// The current parameters
-	boost::shared_ptr<NNParam> nn_param;
+    //The current parameters
+    boost::shared_ptr<NNParam> nn_param;
 
-	void computeThetas(MolData *moldata) override;
+    void computeThetas(MolData *moldata) override;
 };
 
 #endif // __EM_NN_TRAIN_H__
